Bug fix: embedding name error when row_end > n_chunks

+++ tests/search_directory_test.py
@@ -85,16 +85,21 @@ def test_load_after_model_defined(directory_with_embeding_module, embedding_mode
 
 # Test embedding step
 
-variable_names = "start, end, batch, expected_files"
+variable_names = "start, end, batch, clean_files, expected_files"
 values = [
-    (0, None, 100, ["0-76"]),
-    (0, None, 20, ["0-20", '20-40', '40-60', '60-76']),
-    (10, None, 30, ["10-40", "40-70", "70-76"]),
-    (20, 25, 1, ["20-21", '21-22', '22-23', '23-24', '24-25'])
+    (0, None, 100, False, ["0-76"]),
+    (0, None, 100, False, ['0-76']),
+    (10, 40, 10, True, ['0-76']),
+    (0, None, 20, True, ["0-20", '20-40', '40-60', '60-76']),
+    (10, None, 30, True, ["10-40", "40-70", "70-76"]),
+    (20, 25, 1, True, ["20-21", '21-22', '22-23', '23-24', '24-25']),
+    (25, 60, 20, False, ['25-45', '45-60']),
+    (0, None, 15, True, ['0-15', '15-25', '25-45', '45-60', '60-75', '75-76']),
+    (0, 80, 40, True, ['0-40', '40-76'])
 ]
 
 @pytest.mark.parametrize(variable_names, values)
-def test_embedding_creation(directory_with_embeding_module, start, end, batch, expected_files):
+def test_embedding_creation(directory_with_embeding_module, start, end, batch, clean_files, expected_files):
     search = SearchDirectory(directory_with_embeding_module)
     search.embed_text(start, end, batch)
     try:
@@ -102,4 +107,5 @@ def test_embedding_creation(directory_with_embeding_module, start, end, batch, e
             assert f"embeddings ({file}).npy" in os.listdir(directory_with_embeding_module / "embedding_batches")
     finally:
         # remove created files
-        shutil.rmtree(directory_with_embeding_module / "embedding_batches")
+        if clean_files:
+            shutil.rmtree(directory_with_embeding_module / "embedding_batches")
